wrt/tct-ext01-wrt-tests/metadata-support/test_index.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Ge, Zhang <gex.zhang@intel.com>

-->

<html>
  <head>
    <title>WRT Test: metadata-support</title>
    <link rel="author" title="Intel" href="http://www.intel.com"/>
    <link rel="help" href=""/>
    <meta name="flags" content=""/>
    <meta name="assert" content="Check that the web runtime can support the metadata element."/>
    <meta charset="utf-8">
    <script src="resources/testharness.js"></script>
    <script src="resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script language="javascript" type="text/javascript">
       test(function() {
           var metaDataArray = tizen.application.getAppMetaData(null);
           if(metaDataArray.length == 2){
               for(var i=0; i<2; i++){
                   if(metaDataArray[i].key == "key1"){
                       assert_true(metaDataArray[i].value == "", "the web runtime cannot support the metadata element correctly.");
                   }else if(metaDataArray[i].key == "key2"){
                       assert_true(metaDataArray[i].value == "value", "the web runtime cannot support the metadata element correctly.");
                   }else{
                       assert_true(false, "the web runtime cannot support the metadata element correctly.");
                   }
               }
           }else{
               assert_true(false, "the web runtime cannot support the metadata element correctly.");
           }
       }, "Check that the web runtime can support the metadata element");
    </script>
  </body>
</html>
